test: break up the delete nut
