+++ test/nuts/delete/source.nut.ts
@@ -9,7 +9,7 @@ import * as fs from 'node:fs';
 import { fileURLToPath } from 'node:url';
 import * as path from 'node:path';
 import { expect } from 'chai';
-import { execCmd, TestSession } from '@salesforce/cli-plugins-testkit';
+import { execCmd } from '@salesforce/cli-plugins-testkit';
 import { SourceTestkit } from '@salesforce/source-testkit';
 import { FileResponse } from '@salesforce/source-deploy-retrieve';
 import { AuthInfo, Connection } from '@salesforce/core';
@@ -28,62 +28,6 @@ const isNameObsolete = async (username: string, memberType: string, memberName: 
 
   return res.IsNameObsolete;
 };
-
-describe('CustomLabels', () => {
-  let testkit: TestSession;
-
-  before(async () => {
-    testkit = await TestSession.create({
-      project: {
-        gitClone: 'https://github.com/WillieRuemmele/sfdx-delete-customlabel',
-      },
-      scratchOrgs: [{ setDefault: true, config: path.join('config', 'project-scratch-def.json') }],
-      devhubAuthStrategy: 'AUTO',
-    });
-    execCmd('project:deploy:start --source-dir force-app', { ensureExitCode: 0 });
-  });
-
-  after(async () => {
-    await testkit?.clean();
-  });
-  it('will not delete the entire .xml file', () => {
-    const clPath = path.join(
-      testkit.project.dir,
-      'force-app',
-      'main',
-      'default',
-      'labels',
-      'CustomLabels.labels-meta.xml'
-    );
-    const result = execCmd<DeleteSourceJson>(
-      'project:delete:source --json --no-prompt --metadata CustomLabel:DeleteMe',
-      {
-        ensureExitCode: 0,
-      }
-    ).jsonOutput?.result;
-    expect(fs.existsSync(clPath)).to.be.true;
-    expect(fs.readFileSync(clPath, 'utf8')).to.not.contain('<fullName>DeleteMe</fullName>');
-    expect(fs.readFileSync(clPath, 'utf8')).to.contain('<fullName>KeepMe1</fullName>');
-    expect(fs.readFileSync(clPath, 'utf8')).to.contain('<fullName>KeepMe2</fullName>');
-    expect(result?.deletedSource).to.have.length(1);
-  });
-
-  it('will delete the entire .xml file', () => {
-    const clPath = path.join(
-      testkit.project.dir,
-      'force-app',
-      'main',
-      'default',
-      'labels',
-      'CustomLabels.labels-meta.xml'
-    );
-    const result = execCmd<DeleteSourceJson>('project:delete:source --json --no-prompt --metadata CustomLabels', {
-      ensureExitCode: 0,
-    }).jsonOutput?.result;
-    expect(result?.deletedSource).to.have.length(3);
-    expect(fs.existsSync(clPath)).to.be.false;
-  });
-});
 
 describe('project delete source NUTs', () => {
   let testkit: SourceTestkit;
